requirements.txt
requests
beautifulsoup4
python-dotenv
pandas
lxml
